backend-python/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
strawberry-graphql[fastapi]>=0.200.0
python-dateutil>=2.8.0
typing-extensions>=4.0.0
